Use uint32_t instead of size_t according to the spec

+++ include/oneapi/dpl/pstl/hetero/dpcpp/execution_sycl_defs.h
@@ -405,12 +405,12 @@ __kernel_work_group_size(_ExecutionPolicy&& __policy, const sycl::kernel& __kern
 }
 
 template <typename _ExecutionPolicy>
-::std::size_t
+::std::uint32_t
 __kernel_sub_group_size(_ExecutionPolicy&& __policy, const sycl::kernel& __kernel)
 {
     const sycl::device& __device = __policy.queue().get_device();
     const ::std::size_t __wg_size = __kernel_work_group_size(::std::forward<_ExecutionPolicy>(__policy), __kernel);
-    const ::std::size_t __sg_size =
+    const ::std::uint32_t __sg_size =
 #if _USE_KERNEL_DEVICE_SPECIFIC_API
         __kernel.template get_info<sycl::info::kernel_device_specific::max_sub_group_size>(
 #else
